restructured and initial test with Mocap system, need stanleycontroller debugging
