+++ src/svea_charging/package.xml
@@ -3,14 +3,13 @@
 <package format="3">
   <name>svea_charging</name>
   <version>0.0.0</version>
-  <description>TODO: Package description</description>
-  <maintainer email="eliasalbag97@gmail.com">eliasalbag</maintainer>
+  <description>Examples for SVEA</description>
+  <maintainer email="kajarf@kth.se">Kaj Munhoz Arfvidsson</maintainer>
   <license>TODO: License declaration</license>
 
-  <test_depend>ament_copyright</test_depend>
-  <test_depend>ament_flake8</test_depend>
-  <test_depend>ament_pep257</test_depend>
-  <test_depend>python3-pytest</test_depend>
+  <depend>svea_core</depend>
+  <depend>rclpy</depend>
+  
 
   <export>
     <build_type>ament_python</build_type>
