+++ src/svea_charging/scripts/behaviourTree.py
@@ -1,5 +1,5 @@
 import time
-from third_party.btree import NodeStatus, Node, ActionNode, Sequence, Fallback #imported behaviour tree implementation from Elisa Bin's repo: https://github.com/elisabin/btree
+from svea_charging.third_party.btree import NodeStatus, Node, ActionNode, Sequence, Fallback #imported behaviour tree implementation from Elisa Bin's repo: https://github.com/elisabin/btree
