--- src/svea_charging/svea_charging/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/svea_charging
-[install]
-install_scripts=$base/lib/svea_charging
