+++ src/svea_examples/launch/teleop.xml
@@ -1,0 +1,127 @@
+<?xml version="1.0"?>
+<launch>
+    <!-- Launch file arguments -->
+    <arg name="map"             default="floor2"/>
+    <arg name="is_sim"          default="true"/>
+    <arg name="use_foxglove"    default="true"/>
+    <arg name="initial_pose_x"  default="-7.3699"/>
+    <arg name="initial_pose_y"  default="-15.383940368977647"/>
+    <arg name="initial_pose_a"  default="1.0116" /> <!-- wrt to map-->
+    <arg name="map_path" default="$(find-pkg-share svea_core)/maps/$(var map).yaml"/>
+    <arg name="state" default="[-7.4, -15.3, 0.9, 0.0]" />
+
+    <!-- Start Map Server and Foxglove -->
+    <include file="$(find-pkg-share svea_core)/launch/map_and_foxglove.xml">
+        <arg name="map" value="$(var map)" />
+        <arg name="use_foxglove" value="$(var use_foxglove)" />
+    </include>   
+    
+    <!-- floor2 Teleop with amcl -->
+    <group if="$(eval &quot;'$(var map)' == 'floor2'&quot;)">
+         <!-- Real Svea -->
+        <group unless="$(var is_sim)">
+
+            <let name="name" value="self"/>
+
+            <include file="$(find-pkg-share svea_core)/launch/svea.xml">
+                <arg name="is_sim" value="$(var is_sim)"/>
+                <arg name="initial_pose_x" value="$(var initial_pose_x)" />
+                <arg name="initial_pose_y" value="$(var initial_pose_y)" />
+                <arg name="initial_pose_a" value="$(var initial_pose_a)" />
+                <arg name="map" value="$(var map)" />
+            </include>
+
+            <push-ros-namespace namespace="$(var name)"/>
+            
+            <!-- Start teleop_control -->
+            <node name="teleop_control" pkg="svea_examples" exec="teleop_control.py" output="screen">
+                <param name="is_sim" value="$(var is_sim)"/>
+                <param name="localization/base_frame" value="$(var name)/base_link"/>
+            </node>
+
+        </group>
+
+        <!-- SVEA-A -->
+        <group if="$(var is_sim)">
+            <let name="name" value="svea_a"/>
+
+            <include file="$(find-pkg-share svea_core)/launch/svea.xml">
+                <arg name="name" value="$(var name)" />
+                <arg name="is_sim" value="$(var is_sim)"/>
+                <arg name="initial_pose_x" value="$(var initial_pose_x)"/>
+                <arg name="initial_pose_y" value="$(var initial_pose_y)"/>
+                <arg name="initial_pose_a" value="$(var initial_pose_a)"/>
+                <arg name="map" value="$(var map)"/>
+            </include>
+
+            <push-ros-namespace namespace="$(var name)"/>
+
+            <!-- Start teleop_control -->
+            <node name="teleop_control" pkg="svea_examples" exec="teleop_control.py" output="screen">
+                <param name="localization/base_frame" value="$(var name)/base_link"/>
+                <!-- Allow namespace -->
+                <param name="steering_top" value="lli/ctrl/steering"/>
+                <param name="throttle_top" value="lli/ctrl/throttle"/>
+                <param name="highgear_top" value="lli/ctrl/highgear"/>
+                <param name="diff_top" value="lli/ctrl/diff"/>
+            </node>
+        </group>
+    </group>
+
+    
+    <!-- sml Teleop with Mocap-->
+    <group if="$(eval &quot;'$(var map)' == 'sml'&quot;)">
+
+        <group unless="$(var is_sim)">
+            <let name="name" value="self"/>
+            
+            <!-- Start Svea Core -->
+            <include file="$(find-pkg-share svea_core)/launch/svea_mocap.xml">
+                <arg name="name" value="$(var name)" />
+                <arg name="is_sim" value="$(var is_sim)"/>
+                <arg name="initial_pose_x" value="$(var initial_pose_x)"/>
+                <arg name="initial_pose_y" value="$(var initial_pose_y)"/>
+                <arg name="initial_pose_a" value="$(var initial_pose_a)"/>
+                <arg name="map" value="$(var map)"/>
+            </include>     
+            
+            <push-ros-namespace namespace="$(var name)"/>
+
+            <!-- Start teleop_control -->
+            <node name="teleop_control" pkg="svea_examples" exec="teleop_control.py" output="screen">
+                <param name="is_sim" value="$(var is_sim)"/>
+                <param name="localization/base_frame" value="$(var name)/base_link"/>
+            </node>
+        </group>
+
+        <group if="$(var is_sim)">
+            <let name="name" value="svea_a"/>
+
+            <!-- Start Svea Core -->
+            <include file="$(find-pkg-share svea_core)/launch/svea_mocap.xml">
+                <arg name="name" value="$(var name)" />
+                <arg name="is_sim" value="$(var is_sim)"/>
+                <arg name="initial_pose_x" value="$(var initial_pose_x)"/>
+                <arg name="initial_pose_y" value="$(var initial_pose_y)"/>
+                <arg name="initial_pose_a" value="$(var initial_pose_a)"/>
+                <arg name="map" value="$(var map)"/>
+            </include>     
+
+            <push-ros-namespace namespace="$(var name)"/>
+
+            <!-- Start teleop_control -->
+            <node name="teleop_control" pkg="svea_examples" exec="teleop_control.py" output="screen">
+                <param name="is_sim" value="$(var is_sim)"/>
+                <param name="localization/base_frame" value="$(var name)/base_link"/>
+                <!-- Allow namespace -->
+                <param name="steering_top" value="lli/ctrl/steering"/>
+                <param name="throttle_top" value="lli/ctrl/throttle"/>
+                <param name="highgear_top" value="lli/ctrl/highgear"/>
+                <param name="diff_top" value="lli/ctrl/diff"/>
+            </node>
+        </group>
+
+    </group>
+    
+
+</launch>
