+++ src/svea_mocap/scripts/mocap_msg_filter.py
