+++ src/svea_mocap/setup.py
