+++ util/run
@@ -24,6 +24,7 @@ append ARGS \
     -p 8765:8765 \
     -v "$SHARED_VOLUME" \
     -e "TERM=xterm-256color" \
+    --network host \
     --rm
 
 # Enable device access for non-dev / non-macOS hosts
